feat: Issue #490 - 300行制限違反修正完了
8ファイルの300行制限違反解消、Single Responsibility原則実践

分割完了: log_viewer.py, debug_logger.py, doc_classifier.py,
persistence.py, memory_monitor_core.py, optimization_comparison.py,
element_renderer.py, cache_strategies.py

技術的成果:
- 全ファイル300行制限達成
- 全ファイルクラス数制限達成
- 後方互換性維持
- Token効率化実現

Closes #490
