+++ kumihan_formatter/core/rendering/element_renderer.py
@@ -1,84 +1,103 @@
-"""Basic HTML element renderer for Kumihan-Formatter
+"""Element renderer for Kumihan-Formatter - integration of split components
 
-This module handles rendering of basic HTML elements like paragraphs,
-headings, lists, and other simple elements.
+This module provides a unified interface for rendering all HTML elements by
+integrating specialized renderer components.
+
+Issue #490対応 - ファイルサイズ制限遵守により分割統合:
+- basic_element_renderer.py: 基本要素レンダリング
+- heading_renderer.py: 見出し要素レンダリング
+- list_renderer.py: リスト要素レンダリング
+- div_renderer.py: div・details要素レンダリング
 """
 
 from typing import Any
 
 from ..ast_nodes import Node
-from .html_utils import (
-    create_simple_tag,
-    escape_html,
-    process_block_content,
-    process_collapsible_content,
-    process_text_content,
-    render_attributes,
-)
+from .basic_element_renderer import BasicElementRenderer
+from .div_renderer import DivRenderer
+from .heading_renderer import HeadingRenderer
+from .html_utils import create_simple_tag, render_attributes
+from .list_renderer import ListRenderer
 
 
 class ElementRenderer:
-    """Renderer for basic HTML elements"""
+    """Unified element renderer integrating specialized components"""
 
     def __init__(self) -> None:
-        """Initialize element renderer"""
-        self.heading_counter = 0
+        """Initialize element renderer with component instances"""
         self._main_renderer: Any | None = None  # Will be set by main renderer
 
+        # Initialize component renderers
+        self.basic_renderer = BasicElementRenderer()
+        self.heading_renderer = HeadingRenderer()
+        self.list_renderer = ListRenderer()
+        self.div_renderer = DivRenderer()
+
+        # Legacy property for backward compatibility
+        self.heading_counter = 0
+
+    # Basic element methods - delegate to basic_renderer
     def render_paragraph(self, node: Node) -> str:
         """Render paragraph node"""
-        content = self._render_content(node.content, 0)
-        return f"<p>{content}</p>"
+        self.basic_renderer._main_renderer = self._main_renderer
+        return self.basic_renderer.render_paragraph(node)
 
     def render_strong(self, node: Node) -> str:
         """Render strong (bold) node"""
-        content = self._render_content(node.content, 0)
-        return f"<strong>{content}</strong>"
+        self.basic_renderer._main_renderer = self._main_renderer
+        return self.basic_renderer.render_strong(node)
 
     def render_emphasis(self, node: Node) -> str:
         """Render emphasis (italic) node"""
-        content = self._render_content(node.content, 0)
-        return f"<em>{content}</em>"
+        self.basic_renderer._main_renderer = self._main_renderer
+        return self.basic_renderer.render_emphasis(node)
 
+    def render_preformatted(self, node: Node) -> str:
+        """Render preformatted text"""
+        self.basic_renderer._main_renderer = self._main_renderer
+        return self.basic_renderer.render_preformatted(node)
+
+    def render_code(self, node: Node) -> str:
+        """Render inline code"""
+        self.basic_renderer._main_renderer = self._main_renderer
+        return self.basic_renderer.render_code(node)
+
+    def render_image(self, node: Node) -> str:
+        """Render image element"""
+        self.basic_renderer._main_renderer = self._main_renderer
+        return self.basic_renderer.render_image(node)
+
+    def render_error(self, node: Node) -> str:
+        """Render error node"""
+        self.basic_renderer._main_renderer = self._main_renderer
+        return self.basic_renderer.render_error(node)
+
+    def render_toc_placeholder(self, node: Node) -> str:
+        """Render table of contents marker"""
+        self.basic_renderer._main_renderer = self._main_renderer
+        return self.basic_renderer.render_toc_placeholder(node)
+
+    # Div and details methods - delegate to div_renderer
     def render_div(self, node: Node) -> str:
         """Render div node"""
-        content = self._render_div_content(node.content, 0)
-        attributes = render_attributes(node.attributes)
+        self.div_renderer._main_renderer = self._main_renderer
+        return self.div_renderer.render_div(node)
 
-        if attributes:
-            return f"<div {attributes}>{content}</div>"
-        else:
-            return f"<div>{content}</div>"
+    def render_details(self, node: Node) -> str:
+        """Render details/summary element"""
+        self.div_renderer._main_renderer = self._main_renderer
+        return self.div_renderer.render_details(node)
 
+    # Heading methods - delegate to heading_renderer
     def render_heading(self, node: Node, level: int) -> str:
-        """
-        Render heading with ID
+        """Render heading with ID"""
+        self.heading_renderer._main_renderer = self._main_renderer
+        # Sync heading counter
+        self.heading_counter = self.heading_renderer.heading_counter
+        result = self.heading_renderer.render_heading(node, level)
+        self.heading_counter = self.heading_renderer.heading_counter
+        return result
 
-        Args:
-            node: Heading node
-            level: Heading level (1-6)
-
-        Returns:
-            str: HTML heading element
-        """
-        content = self._render_content(node.content, 0)
-
-        # Generate heading ID if not present
-        heading_id = node.get_attribute("id")
-        if not heading_id:
-            self.heading_counter += 1
-            heading_id = f"heading-{self.heading_counter}"
-            node.add_attribute("id", heading_id)
-
-        attributes = render_attributes(node.attributes)
-        tag = f"h{level}"
-
-        if attributes:
-            return f"<{tag} {attributes}>{content}</{tag}>"
-        else:
-            return f"<{tag}>{content}</{tag}>"
-
-    # Heading level methods - delegate to render_heading
     def render_h1(self, node: Node) -> str:
         return self.render_heading(node, 1)
 
@@ -94,207 +113,29 @@ class ElementRenderer:
     def render_h5(self, node: Node) -> str:
         return self.render_heading(node, 5)
 
+    # List methods - delegate to list_renderer
     def render_unordered_list(self, node: Node) -> str:
         """Render unordered list"""
-        items = []
-        if isinstance(node.content, list):
-            for item in node.content:
-                if isinstance(item, Node) and item.type == "li":
-                    items.append(self.render_list_item(item))
-
-        items_html = "\n".join(items)
-        return f"<ul>\n{items_html}\n</ul>"
+        self.list_renderer._main_renderer = self._main_renderer
+        return self.list_renderer.render_unordered_list(node)
 
     def render_ordered_list(self, node: Node) -> str:
         """Render ordered list"""
-        items = []
-        if isinstance(node.content, list):
-            for item in node.content:
-                if isinstance(item, Node) and item.type == "li":
-                    items.append(self.render_list_item(item))
-
-        items_html = "\n".join(items)
-        return f"<ol>\n{items_html}\n</ol>"
+        self.list_renderer._main_renderer = self._main_renderer
+        return self.list_renderer.render_ordered_list(node)
 
     def render_list_item(self, node: Node) -> str:
         """Render list item"""
-        content = self._render_content(node.content, 0)
-        return f"<li>{content}</li>"
-
-    def render_details(self, node: Node) -> str:
-        """Render details/summary element"""
-        # Use special processing for collapsible content to handle lists properly
-        if node.content and isinstance(node.content[0], str):
-            # Process the raw text content with proper list handling
-            content = process_collapsible_content(node.content[0])
-        else:
-            # Fallback to normal content rendering
-            content = self._render_content(node.content, 0)
-
-        summary = node.get_attribute("summary", "詳細を表示")
-
-        # Check if this is a spoiler block
-        is_spoiler = node.get_attribute("spoiler", False) or summary == "ネタバレを表示"
-        class_attr = ' class="spoiler"' if is_spoiler else ""
-
-        # Wrap content in a div to ensure CSS selectors work properly
-        return f'<details{class_attr}><summary>{escape_html(summary)}</summary><div class="details-content">{content}</div></details>'
-
-    def render_preformatted(self, node: Node) -> str:
-        """Render preformatted text"""
-        # For code blocks, preserve whitespace and escape HTML
-        if isinstance(node.content, str):
-            content = escape_html(node.content)
-        else:
-            content = self._render_content(node.content, 0)
-
-        attributes = render_attributes(node.attributes)
-
-        if attributes:
-            return f"<pre {attributes}>{content}</pre>"
-        else:
-            return f"<pre>{content}</pre>"
-
-    def render_code(self, node: Node) -> str:
-        """Render inline code"""
-        if isinstance(node.content, str):
-            content = escape_html(node.content)
-        else:
-            content = self._render_content(node.content, 0)
-
-        attributes = render_attributes(node.attributes)
-
-        if attributes:
-            return f"<code {attributes}>{content}</code>"
-        else:
-            return f"<code>{content}</code>"
-
-    def render_image(self, node: Node) -> str:
-        """Render image element"""
-        filename = node.content if isinstance(node.content, str) else str(node.content)
-        alt_text = node.get_attribute("alt", "")
-
-        # Construct image path
-        src = f"images/{filename}"
-
-        if alt_text:
-            return f'<img src="{escape_html(src)}" alt="{escape_html(alt_text)}" />'
-        else:
-            return f'<img src="{escape_html(src)}" />'
-
-    def render_error(self, node: Node) -> str:
-        """Render error node"""
-        content = escape_html(str(node.content))
-        line_number = node.get_attribute("line")
-
-        error_text = f"[ERROR: {content}]"
-        if line_number:
-            error_text = f"[ERROR (Line {line_number}): {content}]"
-
-        return f'<span style="background-color:#ffe6e6; color:#d32f2f; padding:2px 4px; border-radius:3px;">{error_text}</span>'
-
-    def render_toc_placeholder(self, node: Node) -> str:
-        """Render table of contents marker (should be handled by TOC generator)"""
-        return "<!-- TOC placeholder -->"
+        self.list_renderer._main_renderer = self._main_renderer
+        return self.list_renderer.render_list_item(node)
 
     def render_generic(self, node: Node) -> str:
         """Generic node renderer"""
         tag = node.type
-        content = self._render_content(node.content)
-        attributes = render_attributes(node.attributes)
-
+        content = self.basic_renderer._render_content(node.content)
         return create_simple_tag(tag, content, node.attributes)
-
-    def _render_content(self, content: Any, depth: int = 0) -> str:
-        """
-        Render node content (recursive)
-
-        Args:
-            content: Content to render
-            depth: Current recursion depth
-
-        Returns:
-            str: Rendered content
-        """
-        max_depth = 100  # Prevent infinite recursion
-
-        if depth > max_depth:
-            return "[ERROR: Maximum recursion depth reached]"
-
-        if content is None:
-            return ""
-        elif isinstance(content, str):
-            return process_text_content(content)
-        elif isinstance(content, Node):
-            # Handle single Node objects using main renderer if available
-            if self._main_renderer:
-                return self._main_renderer._render_node_with_depth(content, depth + 1)  # type: ignore
-            else:
-                return f"{{NODE:{content.type}}}"
-        elif isinstance(content, list):
-            parts = []
-            for item in content:
-                if isinstance(item, Node):
-                    # Handle nested nodes using main renderer if available
-                    if self._main_renderer:
-                        parts.append(
-                            self._main_renderer._render_node_with_depth(item, depth + 1)
-                        )
-                    else:
-                        parts.append(f"{{NODE:{item.type}}}")
-                elif isinstance(item, str):
-                    parts.append(process_text_content(item))
-                else:
-                    parts.append(process_text_content(str(item)))
-            return "".join(parts)
-        else:
-            return process_text_content(str(content))
-
-    def _render_div_content(self, content: Any, depth: int = 0) -> str:
-        """
-        Render div content with list marker conversion
-
-        Args:
-            content: Content to render
-            depth: Current recursion depth
-
-        Returns:
-            str: Rendered content with list markers converted
-        """
-        max_depth = 100  # Prevent infinite recursion
-
-        if depth > max_depth:
-            return "[ERROR: Maximum recursion depth reached]"
-
-        if content is None:
-            return ""
-        elif isinstance(content, str):
-            return process_block_content(content)
-        elif isinstance(content, Node):
-            # Handle single Node objects using main renderer if available
-            if self._main_renderer:
-                return self._main_renderer._render_node_with_depth(content, depth + 1)  # type: ignore
-            else:
-                return f"{{NODE:{content.type}}}"
-        elif isinstance(content, list):
-            parts = []
-            for item in content:
-                if isinstance(item, Node):
-                    # Handle nested nodes using main renderer if available
-                    if self._main_renderer:
-                        parts.append(
-                            self._main_renderer._render_node_with_depth(item, depth + 1)
-                        )
-                    else:
-                        parts.append(f"{{NODE:{item.type}}}")
-                elif isinstance(item, str):
-                    parts.append(process_block_content(item))
-                else:
-                    parts.append(process_block_content(str(item)))
-            return "".join(parts)
-        else:
-            return process_block_content(str(content))
 
     def reset_counters(self) -> None:
         """Reset internal counters"""
         self.heading_counter = 0
+        self.heading_renderer.reset_counters()
